+++ memory.c
@@ -1,27 +1,28 @@
 #include "memory.h"
 #include "debug.h"
 #include <string.h>
+#include <stdlib.h>
 
-void* mema(size_t size)
+void* mema(uint64 size)
 {
     void* p = malloc(size);
     return p;
 }
 
-void* mema_zero(size_t size)
+void* mema_zero(uint64 size)
 {
     void* p = mema(size);
     memset(p, 0, size);
     return p;
 }
 
-void* memra(void* cur, size_t size)
+void* memra(void* cur, uint64 size)
 {
     void* p = realloc(cur, size);
     return p;
 }
 
-void* memra_zero(void* cur, size_t size)
+void* memra_zero(void* cur, uint64 size)
 {
     void* p = memra(cur, size);
     memset(p, 0, size);
